extracted AutoAdjuster in saperate module from label.js
